archive/retail-vision-analytics/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "retail-vision-analytics"
version = "1.0.0"
description = "Real-time video analytics for retail environments using NVIDIA DeepStream and TensorRT"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Saurabh", email = "saurabh@example.com"}
]
maintainers = [
    {name = "Saurabh", email = "saurabh@example.com"}
]
keywords = [
    "computer-vision",
    "deep-learning",
    "retail-analytics",
    "nvidia-deepstream",
    "tensorrt",
    "yolov8",
    "object-detection",
    "object-tracking",
    "video-analytics",
    "edge-computing",
    "jetson",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "numpy>=1.24.0",
    "opencv-python-headless>=4.8.0",
    "scipy>=1.11.0",
    "redis>=5.0.0",
    "pyyaml>=6.0.1",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
]
full = [
    "ultralytics>=8.0.200",
    "kafka-python>=2.0.2",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.0",
    "minio>=7.2.0",
    "prometheus-client>=0.19.0",
    "matplotlib>=3.8.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/retail-vision-analytics"
Documentation = "https://github.com/yourusername/retail-vision-analytics#readme"
Repository = "https://github.com/yourusername/retail-vision-analytics"
Issues = "https://github.com/yourusername/retail-vision-analytics/issues"

[project.scripts]
retail-vision = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.txt", "*.md"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | \.venv
    | build
    | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = [".git", ".mypy_cache", ".pytest_cache", ".venv", "build", "dist"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
